apex-predator/requirements.txt
ccxt>=4.3.0
redis>=5.0.0
numpy>=1.26.0
loguru>=0.7.0
aioredis>=2.0.0
python-dotenv>=1.0.0
prometheus-client>=0.20.0
